+++ src/main/java/aws/imgupload/imgupload/service/prod/S3StorageService.java
@@ -17,7 +17,7 @@ public class S3StorageService implements StorageService {
     }
 
     @Override
-    public Optional<byte[]> findImageByIndex(int index) {
+    public Optional<byte[]> findImageByIndex(long index) {
         return Optional.empty();
     }
